add settings and output base for ints
Add settings to the Context, containing base (for ints) and unit for
angle (rad/deg).

New "settings" command shows the current settings.

New "base" command sets the base used to display ints on the
stack. Other types are displayed as before and don't depend on the
base.

Default settings are decimal + radians, as before.